# openapi.yaml
openapi: 3.1.0
info:
  title: Compliance Ops API
  version: 1.0.0
  description: |
    Read-only OSCAL document view plus compliance tracking (frameworks,
    requirements with structured OSCAL control parts and parameters,
    per-statement-part implementation tracking, file and link evidence linked
    to requirements, dashboard, audit).

    OSCAL parameter placeholders (`{{ insert: param, <id> }}`) are resolved at
    import time into NIST-style text: fixed values are joined with `, `;
    selections render as `[Selection: a; b]` or `[Selection (one or more): a; b]`;
    labelled parameters render as `[Assignment: organization-defined <label>]`;
    unknown ids render as `[Assignment: <id>]`. Stored `parts`, `text` and
    `title` never contain placeholders; the parameters themselves are exposed as
    `params` on the requirement detail.

    Statuses use the OSCAL implementation-status vocabulary:
    `implemented`, `partial`, `planned` (default), `alternative`, `not_applicable`.

    Tracking happens per *trackable part*: the items of the control's `statement`
    part (e.g. `ac-2_smt.a`), or the statement itself when it has no items
    (`ac-3_smt`), or a synthetic `<controlId>_smt` when the control has no statement.
    Nested OSCAL controls of class `task` (NIST SP 800-218 SSDF) are not requirements of
    their own: each task is folded into its parent practice as a statement item (id =
...
    control ids only.
...
    requirements.
...

    Conventions:
...
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /v1/oscal/sources:
    get:
      tags: [oscal]
      summary: Discover every configured OSCAL source independently
      description: No query parameters are accepted. Failed sources are unavailable rows and do not fail the response. URLs omit userinfo, query, and fragment.
      responses:
        "200":
          description: Configured sources in operator order
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/OscalSource" } }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/oscal/uploads:
    post:
      tags: [oscal]
      summary: Upload one OSCAL JSON document
      description: Exactly one non-empty multipart `file` is required. Original bytes are limited to 10 MiB and SHA-256 hashed. Arrays, malformed JSON, unsupported roots, and multiple documents are rejected. Filename and media type are ignored.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: Uploaded metadata, content: { application/json: { schema: { $ref: "#/components/schemas/OscalUpload" } } } }
        "400": { $ref: "#/components/responses/InvalidBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { description: Duplicate SHA-256 with existing upload metadata, content: { application/json: { schema: { $ref: "#/components/schemas/OscalUploadConflict" } } } }
        "413": { description: File exceeds 10 MiB, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
    get:
      tags: [oscal]
      summary: List uploaded OSCAL metadata, newest first
      responses:
        "200": { description: Metadata only, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/OscalUpload" } } } } } } }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
  /v1/oscal/uploads/{id}:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [oscal]
      summary: Get upload metadata and stored OSCAL JSON
      responses:
        "200": { description: Upload detail, content: { application/json: { schema: { $ref: "#/components/schemas/OscalUploadDetail" } } } }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
    delete:
      tags: [oscal]
      summary: Delete an upload not linked to an imported framework
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Referenced by an imported framework, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
  /v1/oscal/documents:
    get:
      tags: [oscal]
      summary: List OSCAL document summaries
      description: No query parameters are accepted.
      responses:
        "200":
          description: Document summaries
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/OscalDocumentSummary" }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/OscalUnavailable" }
  /v1/oscal/documents/{documentId}:
    parameters:
      - $ref: "#/components/parameters/documentId"
    get:
      tags: [oscal]
      summary: Get one OSCAL document summary plus the canonical document
      responses:
        "200":
          description: Summary and raw document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OscalDocumentDetail" }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/OscalUnavailable" }
  /v1/oscal/documents/{documentId}/controls:
    parameters:
      - $ref: "#/components/parameters/documentId"
    get:
      tags: [oscal]
      summary: Search controls of a document
      description: Only `q` and `limit` are accepted. `limit` must be an integer in 1..100 (default 50).
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive substring match on control id, title and text.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Matching controls (first `limit`) and the total match count
          content:
            application/json:
              schema:
                type: object
                required: [documentId, items, total]
                properties:
                  documentId: { type: string }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/OscalControl" }
                  total: { type: integer }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/OscalUnavailable" }
  /v1/oscal/documents/{documentId}/controls/{controlId}:
    parameters:
      - $ref: "#/components/parameters/documentId"
      - name: controlId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [oscal]
      summary: Get one control
      responses:
        "200":
          description: Control detail
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    required: [documentId]
                    properties:
                      documentId: { type: string }
                  - $ref: "#/components/schemas/OscalControl"
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/OscalUnavailable" }
  /v1/frameworks/import:
    post:
      tags: [frameworks]
      summary: Import OSCAL catalogs/profiles as frameworks
      description: |
        Imports every `catalog` and `profile` document from all configured OSCAL sources
        (`COMPLIANCE_OSCAL_SOURCE_URL`, comma-separated), or only the document named by
        `documentId`. Profiles are resolved against referenced catalogs available in the
        configured document set or persisted catalog uploads; unresolved or zero-control
        profiles fail with `400` and never create empty frameworks. Re-import is
        intentionally non-destructive: controls present in the source are upserted, while
        controls missing from the new source remain retained snapshots so tracking and
        evidence are preserved. Nothing is deleted (`created` is 0 on the second run).
        Body may be empty.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              maxProperties: 1
              properties:
                documentId:
                  type: string
                  description: OSCAL document uuid. Unknown → 404; non catalog/profile → 400.
                uploadId:
                  type: string
                  description: Persisted upload id. Exactly one of documentId or uploadId may be supplied; only catalog/profile uploads are importable.
      responses:
        "200":
          description: Import results
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ImportResult" }
        "400": { $ref: "#/components/responses/InvalidBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/frameworks:
    get:
      tags: [frameworks]
      summary: List imported frameworks
      responses:
        "200":
          description: Frameworks sorted by title
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Framework" }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
  /v1/frameworks/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [frameworks]
      summary: Get a framework
      responses:
        "200":
          description: Framework
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Framework" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
    patch:
      tags: [frameworks]
      summary: Update the framework short name
      description: |
        Sets the user-editable `shortName` (trimmed, 1..60 characters). Re-imports never
        overwrite a non-empty short name. Unknown fields → 400. Audited as `framework.update`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [shortName]
              properties:
                shortName: { type: string, minLength: 1, maxLength: 60 }
      responses:
        "200":
          description: Updated framework
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Framework" }
        "400": { $ref: "#/components/responses/InvalidBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
  /v1/scope-categories:
    get:
      tags: [scope-categories]
      summary: List reusable scope categories
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive name substring.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        "200":
          description: Categories ordered alphabetically by name.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/ScopeCategory" } }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [scope-categories]
      summary: Create a reusable scope category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScopeCategoryWrite" }
      responses:
        "201":
          description: Created category
          content: { application/json: { schema: { $ref: "#/components/schemas/ScopeCategory" } } }
        "400": { $ref: "#/components/responses/InvalidBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { description: A category with the same case-insensitive name exists. }
  /v1/scope-categories/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    patch:
      tags: [scope-categories]
      summary: Rename a scope category
      description: Optional `expectedUpdatedAt` must exactly match the persisted category timestamp. Omit it for a legacy unconditional rename; `null` is invalid. The check and rename are atomic.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ScopeCategoryPatch" } } }
      responses:
        "200": { description: Updated category, content: { application/json: { schema: { $ref: "#/components/schemas/ScopeCategory" } } } }
        "400": { $ref: "#/components/responses/InvalidBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The optional expectedUpdatedAt is stale, or a category with the same case-insensitive name exists.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    delete:
      tags: [scope-categories]
      summary: Delete an unused scope category
      parameters:
        - name: expectedUpdatedAt
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: Exact current category updatedAt. Omit for legacy unconditional deletion. Invalid or repeated values return 400.
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The expectedUpdatedAt is stale (`resource was modified`) or the category is assigned to a part (`scope category is in use`)., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /v1/requirements:
    get:
      tags: [requirements]
      summary: List requirements
      parameters:
        - name: frameworkId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/RequirementStatus" }
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive substring match on controlId, title and text.
        - name: scopeCategoryId
          in: query
          schema: { type: string }
          description: Repeat or comma-separate UUIDs. ANY category match includes the requirement; unknown UUIDs return zero matches.
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: |
            Page of requirements ordered by framework then controlId. List rows omit
            `parts` (use the detail route); `related` is always present.
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Requirement" }
                  total: { type: integer, description: Total matches ignoring pagination }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
  /v1/requirements/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [requirements]
      summary: Get a requirement with its parts, parameters, per-part tracking and linked evidence
      responses:
        "200":
          description: |
            Requirement detail (`parts`, `params`, `trackableParts` and `evidence` are always
            arrays here). `trackableParts` merges the stored tracking into the statement items in
            document order; a part without stored tracking reads as `planned` with empty owner and
            description and no `dueDate`/`updatedAt`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequirementDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
    patch:
      tags: [requirements]
      summary: Update requirement notes and manual status override
      description: |
        Only the supplied fields change. `statusOverride` accepts a status to set, `null` to
        clear (back to the derived status), or may be omitted. `status`, `owner` and `dueDate` are
        derived from the parts: sending any of them → 400 INVALID_BODY with message
        `status, owner and dueDate are derived from parts`. Unknown fields → 400. Writes one audit
        entry (`requirement.update`, detail `statusOverride {from,to}` and/or `notes: true`).
        Optional `expectedUpdatedAt` must be an RFC3339 timestamp exactly matching the persisted
        requirement `updatedAt`; stale values fail atomically with 409 CONFLICT. Omission preserves
        legacy unconditional writes. `null` is invalid because the requirement always exists.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequirementPatch" }
      responses:
        "200":
          description: Updated requirement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Requirement" }
        "400": { $ref: "#/components/responses/InvalidBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/PreconditionConflict" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
...
  /v1/evidence/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [evidence]
      summary: Get evidence metadata
      responses:
        "200":
          description: Evidence
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Evidence" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
...
  /v1/audit:
    get:
      tags: [audit]
      summary: Recent audit entries, newest first
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEntry" }
        "400": { $ref: "#/components/responses/InvalidQuery" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/TrackingUnavailable" }
...
  /v1/assessments/{uploadId}:
    parameters: [ { name: uploadId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [assessments]
      summary: Get normalized plan, results, or POA&M detail
      responses:
        "200": { description: Normalized assessment with mapping information, content: { application/json: { schema: { $ref: "#/components/schemas/AssessmentDetail" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: One token value from the named set loaded through `COMPLIANCE_API_TOKENS_FILE`. Values must be unique token68-safe strings of at least 12 characters; randomly generated values of at least 43 characters are recommended for production. Token labels are not caller identities and are never sent to the API.
...